float test

+++ registration/spikes_localization_registration/scripts/subtract.py
@@ -36,7 +36,7 @@ g = ap.add_argument_group("Subtraction configuration")
 g.add_argument(
     "--thresholds",
     default=[12, 10, 8, 6, 5, 4],
-    type=lambda x: list(map(int, x.split(","))),
+    type=lambda x: list(map(float, x.split(","))),
 )
 g.add_argument("--nndetect", action="store_true")
 g.add_argument("--dndetect", action="store_true")
